requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
dropbox==12.0.2
garmin-uploader==1.0.10
garminconnect==0.2.25
garth==0.5.2
idna==3.10
lxml==5.3.1
oauthlib==3.2.2
ply==3.11
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
readchar==4.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
stone==3.3.1
typing_extensions==4.12.2
urllib3==2.3.0
withings-sync==4.2.5
